Cleaned up and commented functions: LoadPickupsPositions(), CheckValidRegenSlot(), RandomizPickups(), FilterPickup(), ClearPickupsGot(), KillAllPickups(), CleanUpPickup()
Added critical fix to CheckValidRegenSlot() which resulted in fixing the randomize pickup bug (sometimes 1 of the guns was lost).
It now properly removes the pickup from the invalid regenerate point.
Stopping InitOnePickup() from failing with 'to many pickups of type exist'.

Added alert to CheckValidRegenSlot notifying DEBUG users that a regeneration point falls outside of it's defined group.

Fixed bug introduced in previous alteration to ScatterWeapons().
RegeneratePickup() was never needed because LoseSuperNashram() already added the golden to regenerate list.



git-svn-id: svn://fly.thruhere.net/forsaken@97 1361c5d3-c82f-4d94-b015-8aedb5f389d9